Split service packages to differnt modules for kinesis, route53, iot, cloudsearch
