+++ services/cloudsearch/src/main/resources/codegen-resources/customization.config
