+++ services/cloudsearchdomain/src/main/java/software/amazon/awssdk/services/cloudsearchdomain/SwitchToPostInterceptor.java
