+++ services/cloudsearchdomain/src/main/resources/codegen-resources/codegen.config
