+++ services/cloudsearchdomain/src/main/resources/codegen-resources/customization.config
