+++ services/cloudsearchdomain/src/main/resources/software/amazon/awssdk/services/cloudsearchdomain/execution.interceptors
