+++ services/cloudsearchdomain/src/test/resources/log4j.properties
