+++ services/iot/src/main/resources/codegen-resources/customization.config
