+++ services/iot/src/main/resources/codegen-resources/service-2.json
