+++ services/iotdata/pom.xml
@@ -1,0 +1,36 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+  ~ Copyright 2010-2018 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+  ~  
+  ~ Licensed under the Apache License, Version 2.0 (the "License").
+  ~ You may not use this file except in compliance with the License.
+  ~ A copy of the License is located at
+  ~  
+  ~  http://aws.amazon.com/apache2.0
+  ~  
+  ~ or in the "license" file accompanying this file. This file is distributed
+  ~ on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
+  ~ express or implied. See the License for the specific language governing
+  ~ permissions and limitations under the License.
+  -->
+
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+
+    <parent>
+        <groupId>software.amazon.awssdk</groupId>
+        <artifactId>services</artifactId>
+        <version>2.0.0-preview-11-SNAPSHOT</version>
+    </parent>
+
+    <artifactId>iotdata</artifactId>
+    <name>AWS Java SDK :: Services :: AWS IoT Data</name>
+    <description>The AWS Java SDK for AWS Iot Data Plane Service module holds the client classes that are used for communicating
+        with AWS IoT Data Plane Service
+    </description>
+    <url>https://aws.amazon.com/sdkforjava</url>
+
+
+</project>
