+++ services/iotdata/src/it/java/software/amazon/awssdk/services/iotdata/ServiceIntegrationTest.java
@@ -13,7 +13,7 @@
  * permissions and limitations under the License.
  */
 
-package software.amazon.awssdk.services.iot;
+package software.amazon.awssdk.services.iotdata;
 
 import static org.hamcrest.Matchers.greaterThan;
 import static org.junit.Assert.assertEquals;
