+++ services/iotdata/src/main/resources/codegen-resources/customization.config
