+++ services/iotdata/src/main/resources/codegen-resources/examples-1.json
