+++ services/kinesis/src/main/resources/codegen-resources/codegen.config
