+++ services/kinesis/src/main/resources/codegen-resources/customization.config
