+++ services/kinesis/src/main/resources/codegen-resources/paginators-1.json
