+++ services/kinesis/src/main/resources/codegen-resources/waiters-2.json
