+++ services/kinesisanalytics/src/main/resources/codegen-resources/codegen.config
