+++ services/kinesisanalytics/src/main/resources/codegen-resources/customization.config
