+++ services/kinesisfirehose/src/it/java/software/amazon/awssdk/services/firehose/ServiceIntegrationTest.java
