+++ services/kinesisfirehose/src/main/resources/codegen-resources/codegen.config
