+++ services/kinesisfirehose/src/main/resources/codegen-resources/customization.config
