+++ services/kinesisfirehose/src/main/resources/codegen-resources/service-2.json
