+++ services/route53/src/main/resources/codegen-resources/examples-1.json
