+++ services/route53/src/main/resources/codegen-resources/service-2.json
