+++ services/route53/src/main/resources/codegen-resources/waiters-2.json
