+++ services/route53domains/src/main/resources/codegen-resources/codegen.config
