+++ services/route53domains/src/main/resources/codegen-resources/examples-1.json
