+++ services/route53domains/src/main/resources/codegen-resources/service-2.json
